Add Baeyer-Villiger kinetics families
The Baeyer-Villiger reaction occurs in two steps. The first step is
typically catalyzed by an acid, so there is only one family for the
first step. The second step is usually uncatalyzed if it occurs with
a peroxy acid, but has to be catalyzed if it occurs with a pure
hydroperoxide.